Added functions to add offine test marks

+++ modules/manage_test/function/delete_test.php
@@ -8,13 +8,16 @@ $dbConnect = new DBConnect(Constants::SERVER_NAME,
     Constants::DB_NAME);
 $teacher_id = 1; //To Do: Change This
 $test = new Test($dbConnect->getInstance());
-if($test->deleteTest($_REQUEST["id"],$teacher_id)){
-    $test->parentPageRedirect("Test deleted successfully");
-    if(file_exists("offline_test/" . $_REQUEST["id"] . ".pdf")){
-        unlink("offline_test/" . $_REQUEST["id"] . ".pdf");
+if(isset($_REQUEST["id"])){
+    if($test->deleteTest($_REQUEST["id"],$teacher_id)){
+        $test->parentPageRedirect("Test deleted successfully");
+        if(file_exists("offline_test/" . $_REQUEST["id"] . ".pdf")){
+            unlink("offline_test/" . $_REQUEST["id"] . ".pdf");
+        }else{
+        }
     }else{
+        $test->parentPageRedirect("Error while deleting test");
     }
 }else{
     $test->parentPageRedirect("Error while deleting test");
 }
-?>
